move code to Spark 3.5
