+++ extensions/spark/kyuubi-extension-spark-3-4/pom.xml
@@ -67,6 +67,8 @@
             <groupId>org.apache.kyuubi</groupId>
             <artifactId>kyuubi-util-scala_${scala.binary.version}</artifactId>
             <version>${project.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
